Update progress
Beberapa perubahan untuk apk :
- fix dialog yang kurang pada halaman login jika tidak mengisi kolom
- penggunaan shared preferences pada login, sehingga ketika sudah login, maka setiap menjalankan aplikasi akan tetap memunculkan halaman dashboard kecuali melakukan logout yang tersedia pada menu drawer
- disabled tombol back bawaan pada perangkat, sehingga tombol back tidak bisa digunakan selama aplikasi dijalankan, dan navigasi dilakukan menggunakan tombol yang tersedia
- pada finished patient, ketika masuk ke halaman detail tidak akan melakukan update (hanya dapat melihat detail informasi saja termasuk diagnosis dan prescription)
- pada halaman update outpatient, jika hanya menginput spasi atau tidak menginput sama sekali pada kolom diagnosis dan prescription, ketika diklik tombol update akan memunculkan warning atau peringatan untuk mengisi kolom terlebih dahulu.
