+++ capstone_project_hospital_management/.vscode/settings.json
@@ -27,6 +27,7 @@
         "Pria",
         "proccessing",
         "savaing",
+        "Scarlett",
         "sqflite",
         "sudah",
         "tanggal",
