+++ capstone_project_hospital_management/lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:capstone_project_hospital_management/presentation/splash/splash_screen.dart';
+import 'package:capstone_project_hospital_management/screen/splash/splash_screen.dart';
 import 'package:flutter/material.dart';
 
 void main() {
