+++ capstone_project_hospital_management/lib/widget/from_API/patient_single_list_done_api.dart
@@ -1,5 +1,6 @@
 import 'package:capstone_project_hospital_management/model/outpatient_model.dart';
-import 'package:capstone_project_hospital_management/presentation/patient/detail/patient_detail_api.dart';
+import 'package:capstone_project_hospital_management/screen/patient/detail/patient_detail_api.dart';
+import 'package:capstone_project_hospital_management/screen/patient/detail/patient_detail_api_done.dart';
 import 'package:capstone_project_hospital_management/widget/settings.dart';
 import 'package:flutter/material.dart';
 
@@ -83,8 +84,8 @@ class PatientSingleListAPIDone extends StatelessWidget {
               child: IconButton(
                 onPressed: () {
                   Navigator.of(context)
-                      .pushReplacement(MaterialPageRoute(builder: (context) {
-                    return PatientDetailPageAPI(
+                      .push(MaterialPageRoute(builder: (context) {
+                    return PatientDetailPageDoneAPI(
                       patient: patient,
                     );
                   }));
